Use run script to orchestrate the whole thing
- Run bpftrace (or whatever) in the background, and stop it after the
  benchmark.
- Copy all the results into this repo so we can save them.

This requires collecting the sudo password ourselves, since the
benchmark sometimes takes long enough that sudo wants us to type it
again later.

I hit lots of issues with running sudo in the background messing up our
terminal. It seems like `sudo --background` is the only way around it.
But then the process is no long a child, so signalling it is more
complicated. This would be a lot easier if bpftrace let us just use
`CAP_BPF` instead of being uid 0.

Now I've collected results with 1% invalid data from all foreign key
implementations. It's nice to have it automated.